Some tweaks for beta

+++ src-tauri/yaak_plugin_runtime/src/events.rs
@@ -364,7 +364,7 @@ pub struct GetHttpRequestByIdResponse {
 #[ts(export)]
 pub struct FindHttpResponsesRequest {
     pub request_id: String,
-    pub limit: Option<i64>,
+    pub limit: Option<i32>,
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, TS)]
